Support resolution of unencrypted code from plaintext asar file (closes #41)
